FIX: KM 단위 이하의 거리 계산이 표시 안되는 것을 수정
dist 는 일과 해결사의 거리차를 KM 단위로 보여주는 dto 멤버 였음. int였던 dist를 double로 바꿔서 KM 이하 계산도 표시 하도록 함.

+++ src/main/java/spot/spot/domain/job/dto/response/NearByJobResponse.java
@@ -22,7 +22,7 @@ public class NearByJobResponse {
     @Schema(description = "돈")
     private int money;
     @Schema(description = "일과 현 사용자가 보낸 위치간의 거리 차이(km 단위)")
-    private int dist;
+    private double dist;
     @Schema(description = "카카오페이 결제 번호")
     private int tid;
 }
